CSSMS3 OSBs working with OSBArray; tryin AceButton next...

+++ MRSRCPCM/src/LocalDisplay.cpp
@@ -409,7 +409,7 @@ LocalDisplayClass::Pages LocalDisplayClass::GetCurrentPage()
 	return currentPage;
 }
 
-void LocalDisplayClass::PrevPage(byte /*value*/)
+void LocalDisplayClass::PrevPage(int /*value*/)
 {
 	//TODO: Untested
 	if (currentPage <= LocalDisplayClass::Pages::SYS)
@@ -424,7 +424,7 @@ void LocalDisplayClass::PrevPage(byte /*value*/)
 	}
 }
 
-void LocalDisplayClass::NextPage(byte /*value*/)
+void LocalDisplayClass::NextPage(int /*value*/)
 {
 	if (currentPage >= LocalDisplayClass::Pages::NONE)
 	{
@@ -471,7 +471,7 @@ TFT_eSPI* LocalDisplayClass::GetTFT()
 	return &tft;
 }
 
-void LocalDisplayClass::SetDisplayBrightness(byte brightness)
+void LocalDisplayClass::SetDisplayBrightness(int brightness)
 {
 	Brightness = brightness;
 	analogWrite(LCD_BL, Brightness);
